backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic-settings
python-multipart
python-dotenv
google-generativeai
pypdf
typing-extensions
